ui (order change): move DisplayAdapter to the top
